requirements.txt
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2
pyyaml==6.0.1

catboost
xgboost
lightgbm

fastapi==0.104.1
uvicorn[standard]==0.24.0

pydantic==2.5.0
pydantic-settings==2.1.0

sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9

python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

slowapi==0.1.9
redis==5.0.1

httpx==0.24.1
email-validator
python-dotenv==1.0.0

asyncpg==0.29.0
supabase==2.3.0
